mysite/requirements.txt
django
djangorestframework
environs
social-auth-core
social-auth-app-django

asgiref
django-extensions
pillow
sqlparse
tzdata

djoser
django-cors-headers
djangorestframework-simplejwt
dj-rest-auth[with_social] 
django-allauth

requests

sentence-transformers 
scikit-learn




requests

sentence-transformers
scikit-learn

#pentru scraping:
selenium
webdriver-manager
beautifulsoup4

